Exercises for 4d

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
@@ -10,13 +8,11 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const user = request.user
 
-  if (!request.token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
+  if (!user) {
+    return response.status(401).json({ error: 'token missing' })
   }
-
-  const user = await User.findById(decodedToken.id)
 
   if (!body.title && !body.url) {
     return response.status(400).json({ error: 'Missing title and url' })
@@ -38,8 +34,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  console.log(user)
+
+  if (blog.user._id.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    user.blogs = user.blogs.filter(b => b.toString() !== blog._id.toString())
+    await user.save()
+    response.status(204).end()
+  } else {
+    return response.status(400).json({ error: 'Cannot delete blog you have not created' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
